Create application-layer.json
This code creates the 2 Lambda functions that we created the custom resources for, along with permissions for Lambda to carry out these functions.
And an API Gateway with 2 endpoints: User Input Endpoint to accept and process the information user provide and Order Endpoint to handle and process user orders. 
And each endpoint has 2 main methods setup: OPTIONS method to describe the communication options for the target resource and POST method to allow users to send data to the endpoint.